backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
fastapi==0.135.3
h11==0.16.0
idna==3.11
paho-mqtt==2.1.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
slowapi==0.1.9
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
alembic==1.15.2
chirpstack-api>=4.7.0
grpcio>=1.60.0
sqlalchemy==2.0.41
uvicorn==0.44.0
websockets==15.0.1
